requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
docopt==0.6.2
enum34==1.1.6
idna==2.8
isort==4.3.4
kiwisolver==1.0.1
lxml==4.3.3
matplotlib==3.0.3
numpy==1.16.3
pep8==1.7.1
Pillow==6.0.0
pipreqs==0.4.9
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.24.2
vk==2.0.2
vk-api==11.4.0
xlrd==1.2.0
yarg==0.1.9
gunicorn==19.9.0